Entities almost done

+++ ProjectDefense.Data/Scripts/MainScripts/0014_create_table_attribute_options.sql
@@ -2,7 +2,7 @@ create table attribute_options(
     id            serial primary key,
     attribute_id  int not null references attributes(id) on delete cascade,
     label         varchar(100) not null,
-    sort_order    int not null default 0,
+    sort_order    smallint not null default 0,
     unique (attribute_id, label),
 
     created_user_id     uuid not null,
